use solid in rightSidebarTexts
